exam-grader/backend/requirements.txt

fastapi>=0.110
uvicorn[standard]>=0.27

pydantic>=2.6
pydantic-settings>=2.2
python-dotenv>=1.0

httpx>=0.27
tenacity>=8.2

supabase>=2.4

python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
bcrypt>=4.1
python-multipart>=0.0.9

Pillow>=10.2
paddleocr>=2.7
numpy>=1.26
pymupdf>=1.24

cloudinary>=1.39

pandas>=2.2
openpyxl>=3.1
reportlab>=4.1

anthropic>=0.25
openai>=1.12

pytest>=8.1
pytest-asyncio>=0.23
